refactor(member): 라우팅 파일 프로젝트 구조 변경
